tests/requirements.txt
pytest>=7.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
numpy>=1.20.0